Issue #2742205 by smccabe: Port commerce_cart tests to BrowserTest
